modified emailer view to view event reminders properly
I have modified emailer view to always print out the email template's
content instead of having the option of printing an element view, which
was used when the event reminder was a standard text that cannot be
modified.

I have fixed saving event reminders to also save the email template's
subject line, which was not done in the previous commit.

Lastly, I removed the eventReminder email element, since it is no
longer in use.