Fix #564, students unable to submit 0 mark in the mixevaL
When select 0 mark, system shows require field are not completed. When
select a mark other than 0, the evaluation can be submitted, however,
when changing the mark to 0 and after saving, it changed back to the
original value.

The LOM should start with 1. Zeros are all for comments questions.
The LOM values are inconsistent between evaluations with and without
zero marks. E.g. with zero marks, LOM starts with 0, and without zero
marks, LOM starts with 1.