Added the ability to un/release specific comments for rubric evaluations
This is the first commit for allowing instructors to un/release specific
comments of a rubric evaluation (similar to what can be currently done
with rubric evaluations).

Included in this commit is:
* modifications to the sql scripts (include upgrade scripts) to include
an attribute to mark whether each specific comment is released
* buttons to "Save changes" of releasing the selected comments
* buttons to "Un/release all" comments for an evaluatee or the whole
group
* internal check to mark a group_event as reviewed if all of the
evaluation_details in the group has been modified after the close
date of the evaluation - in other words, the detail was marked as
released or unreleased using Save Changes or un/release all (or
combination of both).

Along with this additional feature we have remodelled the process of
un/releasing evaluation results, which is described below. We also
took into account the possibility of an instructor changing the
auto-release attribute BEFORE and AFTER the close date of the event.
* BEFORE close date
   * grades/comments release links/buttons would not be visible, to
     remove the ambiguity of whether the results have been reviewed
   * if auto-release is turned on, evaluation details are defaulted
     to have comment_release/grade-release=1
   	* they are set to zero if auto-release if off
   * if the auto-release status is changed during this period, all
     submitted evaluations will be marked appropriately
	* eg. from off to on - mark all release status to 1

* AFTER close date
   * grades/comments release links/buttons would be visible
   * if auto-release is on, students can see the results as soon
     as the evalution results start date has been reached
   * if the auto-release status is changed during this  period, all
     submitted evaluations will be marked appropriately IF the
     group event is not marked as "reviewed".
	* if the group_event is marked as reviewed, no changes to
          the release status of the details would occur, because
          we assume the instructor has reviewed the comments and would
          like to keep those changes.