modified student view of rubric evaluation results
The student view of the results for rubric per evaluations has been
modified to reflect the individual releases of comments that were
made possible on the instructor's view of the same results.

To make the view more readable:
- I have made the grades (including titles) not visible if they are
not released. Reason behind this decision is if we couple that with
some comments being unrelease, we will have a long list of "n/a", which
discourages students to read the results.
- I have also made the student view of their own evaluation grades/comments
for their team members to be visible whether the grades or comments have been
released or not. This is the same behaviour from before.
    - consequently, comments are not filtered in this section, I'm assuming
      the student knows what they wrote (whether it is good or bad)