implemented the feature to un/release individual comments in mixeval
- now users can un/release individual comments; so that inappropriate
comments can be hidden
- fixed the broken tests in the evaluation component test suite; some
were broken due to the change in data format made for the new feature
- moved un/releasing all grades and comments for an evaluatee or the
whole group buttons near the top of the detailed evaluation results
page. It is placed in a table format. This allows users to easily
release groups of grades and comments without having to navigate
through multiple accordions. The second reason is the user can now
easily see the release status of grades and comments for any students
in the group without the need to scroll through the results view.