updated the auto_release mechanism for edit events
Now, if a user changes the auto release status of a event, the release
status of the submitted evaluations are automatically change. The only
exception is if the group event is already marked as reviewed; then the
release status are left untouched. I've also removed the messages alerting
the user that the event's result is auto_release since the system now
changes the release status automatically. Lastly, I fixed the checkboxes
for un/releasing individual comments for rubric and mixed evaluations from
showing up when the event has not ended (the "save changes" and "un/release"
buttons are already not visible when the event has not ended).