First commit for  #537, Soft Delete Users
For this feature, I have used the existing "record_status" attribute
for users. If the user is active, they have a "A" status; meanwhile
a user that is inactive will have a "I" status.

When a user is "deleted", their status flips from "A" to "I" and all
of their records in user_enrol, user_tutor, user_course, group_members,
survey_group_members are simply deleted, because the effect can be
easily reversed.

In the users index, users that are inactive will no longer show up.

When a user is trying to add a new user with an inactive username, an
option is given to simply readd the user. Also during user import,
if a user that is being enrolled is inactive, their status is simply
flipped back to active. Similar implementation is also done for importing
users via the API

Lastly, inactive users are denied access to the system. The error
message currently does not appear when logging in with the
default module however. A ticket has been created to address this
issue (Ticket #583).