Fix #578: Revise event API to return all available events
I have modified the userEvents API to return all events for the user.
Moreover, I have included two optional filters that can be used to
narrow down the list, which are available for submissions and results
released.

The original url such as - /v1/users/:username/events still works but
returns all events instead of events that are available for submissions.

The new urls are:
- /v1/courses/:courseId/users/:username/events/sub/0/results/0, for returning events
in specified course that match the filters
- /v1/users/:username/events/sub/0/results/0, for returning events for
user that matches the filters.

Filter 'sub' - if sub is 1, only events that are available for submission
is returned
Filter 'results' - if results is 1, only events that have results released
are returned
If both filters are one - events that matches one of the two criteria
will be returned.