fixed super admins unable to view all instructors and faculty admins
The reason was super admins were also using the custom filter for
faculty admins; which means they can only see instructors and admins
that are in the faculties the super admin is enroled in. This has been
fixed by fixing the condition to skip the filtering for super admins.

I have also modified the test data to not enrol the super admin in the
two test faculties since in production super admins are not automatically
enroled in all faculties (and departments)

Due to the enrolment removal mentioned above, the "getIdsByUserId"
function test became broken when a super admin user id is passed.
There is no easy fix to this until we can check for a user's permission
based on the userId passed in (instead of the logged in user). Therefore,
I have commented that one test case untilk I have more time to implement
a permanent fix for this.