implemented some validations to event edit
1. If the form is not fully loaded, and the user tries to submit
the form, an error message will appear. This is so that unwanted
modifications to the event will not be saved.

2. Another validation has been added to ask the user for confirmation
that they wanted to remove groups that were initially in the event.
If no original groups have been removed, no confirmation dialog will
appear when the form is submitted. If one or more original group has
been removed, a confirmation dialog will appear. Whether new groups
have been added to the event will not affect conditions.

3. I've also updated the events system test to expect the confirmation
dialog and to click accept before continuing with the rest of the tests.