Fix emails templates filling in wrong names.
Users were getting emails where they were being addressed by the wrong first
name. Turns out that the email template replacement was modifying the original
template, so it replaced template variables with the info of the first user and
alone after that. Everyone was getting same email as the first user.