fixed manual releasing mixed evaluation comments
The comment release status on the student view is using the
comment_release status attribute for evaluation_mixevals, which
is currently no longer in use - due to the implementation of the
ability to release individual comments

Tests done:
- Mixed
	- manually releasing all comments
	- manually releasing selected comments
	- not releasing any comments
	- auto-releasing all comments
	- auto-releasing comments / unrelease selected comments
- Rubrics - manual and auto release comments (all/selected)
- Simple - manual and auto release comments (all/selected)

No change is needed for Rubrics and Simple because the comment_release
status in their respective evaluation object is still in use. I've also
checked that releasing grades is still functional. Since releasing
grades is all or nothing, we do not need to change the computation
of the grade release boolean to array_sum.