fixed self-evaluation features showing up even though the feature
was disabled for the events

- in the case of changing the self-evaluation option after the event
has opened, the submitted self-evaluations will not be included in
the evaluation results UI for instructors and students, CSV export,
and PDF export
	- the self-evaluation results will appear again when the
	option is re-enabled
	- scores would also not include the self-evaluation submissions
	when self-evaluation is disabled
- for mixed evaluation, if self-evaluation is disabled both the peer
evaluation questions for the evaluator to evaluate themselves and
the self-evaluation specific questions will not appear.
- mixed evaluation template view will still disable the self-evaluation
specific questions
- mixed evaluation template preview (from event form), will not display
self-evaluation specific questions if the feature is diabled.