UPGRADE php to 7.4, composer to 2.4
Set composer.json to target php 7.4 and ran composer update.

Updated Dockerfile and Dockerfile-app-unittest to use the php 7.4 image.

Dockerfile-app-unittest needed to update the oauth package to latest
2.0.7 or it'll fail to compile for php 7.4.

Also updated the way composer is installed by the Dockerfiles, we just
copy the binary from the official composer image. Along with this, we
updated composer from 1.8 to 2.4.