UPDATE composer to target php8.0, rm phpdocumentor
Set composer to target php8.0 and then ran composer update.

Removed phpdocumentor as it was bringing in zendframework dependencies
that weren't compatible with php8. Found out that phpdocumentor's docs
recommends running it as a standalone tool rather than installing via
composer anyways. Checking build.xml revealed that phpdocumentor was
removed from 'build-ci' for breaking with php7, so it seems like we
might not have been using it for a while?