UPGRADE phpexcel to phpspreadsheet
PHPExcel is now an archived project and they recommend switching to
PhpSpreadsheet, the successor to PHPExcel. The recommended upgrade path
is using an automated tool called RectorPHP. I ran Rector and the
changes seem reasonable, but not sure if it's working, but at least it
doesn't throw any exceptions. Rector also made that other changes that
I've also checked in to this commit.

Looking at it closer, it seems that we only have 2 export types: csv and
pdf in the UI. Guessing the Excel option was removed at some point? So
I'm not sure if this was necessary, maybe we could've just removed the
entire feature.

PhpSpreadsheet is brought in by composer, so we needed to autoload it.
I've modified app/webroot/index.php so that it'll require necessary the
composer autoload.php file.

PhpSpreadsheet requires the zip extension, this was added to the
Dockerfiles.