ADD export csv component test
Simple test for the export csv component. It's not quite a complete
test, as the component writes the csv to php://output and I couldn't
figure out how to get intercept the stream. So it really only checks
that the data was actually gathered correctly.

The test gets the array that's fed to fputcsv that generates the actual
csv file, so we have to reduplicate the fputcsv part of the component
and write it to a string this time so the test can compare the expected
and actual output.