FIX broken build due to imsglobal/caliper
imsglobal/caliper repo is no longer publicly accessible. I've previously
fixed this for the wiki before so I've just applied the same fix, which
is basically just installing the lib from a zip of the version we need.

Also updated to php 8.2 and compose 2.6.

Made a note for the password for the sample data users.