EDIT docker update and cleanup
Composer 2.7 is out, instead of pinning our version to a specific minor
version, it's probably safe to just pin our version to major version 2.

composer's '--no-suggest' option is deprecated and does nothing now, so
it was removed.

Updated mariadb to 10.11 LTS.

Added adminer as a front-end UI for mariadb to help with dev.

Removed zend extension opcache.so as the version is very old and doesn't
seem necessary.

Removed container names as docker compose best practice is not to use
them.