FIX missing cake_sessions table on new install
The cake_sessions table is used by the worker, on a fresh install, the
worker will crash due to it missing. I've added it to the database
create sql to fix this. Originally, it was meant to be created by
app/config/bootstrap.php as a sign of successful install, but the worker
gets up and running before that gets to run.

Fix a '>' appearing at the top of the page if no Google analytics is
confirmed. This is an extra > in the Google analytics script block.

DJJob was still being pulled in by git submodule, looking at the
submodules, we don't use most of it anymore, so I added a manual clone
to the Dockerfile to pull it in.