FIX mixeval release all comments timeout
1400 student course, instructor reports "Release All Comments" reaches
timeout. They have to keep retrying until all comments are released.

This was for a mixeval. Checking the codepath and CakePHP 1.3 docs, I
realized that this was due to CakePHP eager loading all relations. The
queries became *very* expensive since we're loading 1400 students worth
of submissions every time it processes 1 submission. And each submission
is unfortunately processed 1 by 1, so it's loading 1400 submissions 1400
times (at least).

The fix is to convert the queries to use the find() call where you can
pass in the 'recursive' parameter to set it to -1, so it doesn't load
relations. I've done this only for mixevals, as the course seems to only
use mixevals.

One of them uses a method on Event. So I modified the method to take a
lazy param that defaults to false and only adds the recursive param if
lazy is true. This preserves the original behaviour in case some code
depends on the original eager loading behaviour.