FIX github action build fail due to djjob clone
This djjob clone was a fix for local builds where I wasn't pulling in
submodules. In the github action build, the checkout action is
automatically configured to pull in submodules, so the build fails when
trying to clone the djjob repo since it already exists.

For now, I'll remove the manual djjob clone.